fix: Mudanças nos models e migrations

+++ back-end/src/database/migrations/20220531191852-create-procuts.js
@@ -14,14 +14,12 @@ module.exports = {
         unique: true,
       },
       price: {
-        type: Sequelize.FLOAT(4,2),
+        type: Sequelize.DECIMAL(4,2),
         allowNull: false,
-
       },
       url_image: {
         type: Sequelize.STRING(200),
         allowNull: false,
-
       }
     });
   },
